+++ src/app/write/page.tsx
@@ -3,7 +3,7 @@ import React, { Suspense } from 'react';
 
 export default function WritePage() {
   const Editor = React.lazy(
-    async () => await import('@/app/components/WriteComponent'),
+    async () => await import('@/components/WriteComponent'),
   );
   return (
     <Suspense fallback={<div>...loading</div>}>
