+++ src/components/Login.tsx
@@ -1,7 +1,7 @@
 'use client';
 import React, { useEffect, useState } from 'react';
 import { signInWithEmailAndPassword, signOut } from 'firebase/auth';
-import { auth } from '../../../firebase';
+import { auth } from '../../firebase';
 
 export default function Login() {
   const [information, setInformation] = useState({
