+++ src/components/WriteComponent.tsx
@@ -3,7 +3,7 @@ import dynamic from 'next/dynamic';
 
 export default function WrtieComponent() {
   const Editor = dynamic(
-    async () => await import('@/app/components/Custom-Editor'),
+    async () => await import('@/components/Custom-Editor'),
     {
       ssr: false,
     },
