+++ src/context/AuthContext.tsx
@@ -1,6 +1,6 @@
 'use client';
 import React, { createContext, useEffect, useState } from 'react';
-import { auth } from '../../../firebase';
+import { auth } from '../../firebase';
 
 export const CreateAuthContext = createContext<any>(null);
