+++ src/utils/UploaImage.ts
@@ -3,7 +3,7 @@ import {
   uploadBytes,
   getDownloadURL,
 } from 'firebase/storage';
-import { storage } from '../../../firebase';
+import { storage } from '../../firebase';
 
 export class CustomUploadAdapter {
   loader: any;
